+++ pkg/probodconfig/slack_config.go
@@ -1,4 +1,4 @@
-// Copyright (c) 2025-2026 Probo Inc <hello@getprobo.com>.
+// Copyright (c) 2026 Probo Inc <hello@getprobo.com>.
 //
 // Permission to use, copy, modify, and/or distribute this software for any
 // purpose with or without fee is hereby granted, provided that the above
@@ -12,18 +12,9 @@
 // OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR
 // PERFORMANCE OF THIS SOFTWARE.
 
-package probod
+package probodconfig
 
-type MailerConfig struct {
-	MailerInterval int        `json:"mailer-interval"`
-	SenderName     string     `json:"sender-name"`
-	SenderEmail    string     `json:"sender-email"`
-	SMTP           SMTPConfig `json:"smtp"`
-}
-
-type SMTPConfig struct {
-	Addr        string `json:"addr"`
-	User        string `json:"user"`
-	Password    string `json:"password"`
-	TLSRequired bool   `json:"tls-required"`
+type SlackConfig struct {
+	SenderInterval int    `json:"sender-interval"`
+	SigningSecret  string `json:"signing-secret"`
 }
